+++ source/core.py
@@ -314,7 +314,7 @@ def resetConfiguration(factoryDefaults=False):
 	log.debug("terminating tones")
 	tones.terminate()
 	log.debug("terminating sound split")
-	audio.soundSplit.terminate()
+	audio.terminate()
 	log.debug("Terminating background braille display detection")
 	bdDetect.terminate()
 	log.debug("Terminating background i/o")
@@ -347,7 +347,7 @@ def resetConfiguration(factoryDefaults=False):
 	tones.initialize()
 	# Sound split
 	log.debug("initializing sound split")
-	audio.soundSplit.initialize()
+	audio.initialize()
 	# Speech
 	log.debug("initializing speech")
 	speech.initialize()
@@ -739,7 +739,7 @@ def main():
 	log.debug("Initializing sound split")
 	import audio
 
-	audio.soundSplit.initialize()
+	audio.initialize()
 	import speechDictHandler
 
 	log.debug("Speech Dictionary processing")